# specs/swaggerdoc.yaml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Medten Product API'
  description: 'Interact with the Medten product database using this API.'
servers:
  - url: https://mgtwitxxi1.execute-api.ap-southeast-1.amazonaws.com/Prod/api/v1/
    description: API Test Environment
  - url: https://v73kddyeqk.execute-api.ap-southeast-1.amazonaws.com/Prod/api/v1/
    description: API POC (Proof of Concept) Environment
  - url: http://localhost:3000/api/v1/
    description: Developer Local Test Environments
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/gaumeister/medten-product-api/1.0.0
paths: 
...
  /products/${id}/equipment:
    summary: The equipment for a product.
    parameters:
    - name: id
      in: path
      description: the product id
      required: true
      schema:
        type: integer
    get:
      summary: gets all equipment for a product
      tags:
        - products
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  product_equipment_connect_views:
                    type: array
                    items:  
                      $ref: '#/components/schemas/ProductEquipmentView'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
    post:
      summary: replaces the current set of equipment for a product
      description: Replaces the current set of equipment for a product, adding those that need to be added, updating any existing, and removing any that no longer match the payload. Equipment items are differentiated by the `equipment_id` property.
      tags:
        - products
      requestBody:
        description: >
          The new set of equipment for the product
        required: true
        content:
          application/json:
            schema:
              type: array
              items:  
                $ref: '#/components/schemas/ProductEquipment'
                  
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:  
                  $ref: '#/components/schemas/ProductEquipment'
              
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
...
  /equipment:
    get:
      summary: searches for equipment
      tags:
        - equipment
      parameters:
        - in: query
          name: search_term
          schema:
            type: string
          description: Using the `search_term` parameter searches the `model`, `type_en`, `type_zh`, `brand_en`, and `brand_zh` properties using a wildcard search. **Using this parameter overrides any other query parameters (i.e. other query parameters will be ignored).**
        - in: query
          name: brand_id
          schema:
            type: integer
        - in: query
          name: brand_en
          schema:
            type: string
        - in: query
          name: brand_zh
          schema:
            type: string
        - in: query
          name: created
          schema:
            type: string
          description: (ISO-8601 date format)
        - in: query
          name: id
          description: the equipment id
          schema:
            type: integer
        - in: query
          name: model
          schema:
            type: string
        - in: query
          name: type_en
          schema:
            type: string
          description: the name of the equipment type (English)
        - in: query
          name: type_zh
          schema:
            type: string
          description: the name of the equipment type (Chinese)
        - in: query
          name: updated
          schema:
            type: string
          description: (ISO-8601 date format)
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  equipment:
                    type: array
                    items:
                      $ref: '#/components/schemas/EquipmentView'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
          
    post:
      summary: creates a new equipment item
      tags:
        - equipment
      requestBody:
        description: >
          The equipment to be created. Note the `id` property should be excluded. It will be returned in the response.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipment'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Equipment'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'  
  /equipment/${id}:
    summary: Represents equipment.
    description: An equipment is any entity having a model number that could potentially be assigned to a product.
    parameters:
    - name: id
      in: path
      description: the equipment id
      required: true
      schema:
        type: integer
    get:
      summary: gets equipment by id
      tags:
        - equipment
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Equipment'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
    put:
      summary: updates an equipment item
      tags:
        - equipment
      requestBody:
        description: >
          The equipment item to be updated.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Equipment'
      responses:
        200:
          description: returns the equipment, including extended information
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/Equipment'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /equipment/${id}/available_regions:
    parameters:
    - name: id
      in: path
      description: the equipment id
      required: true
      schema:
        type: integer
    get:
      summary: gets available regions for a given equipment item
      tags:
        - equipment
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/EquipmentAvailableRegionView'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
    post:
      summary: replaces the available regions for an equipment item
      description: Replaces the available regions for an equipment item, adding those that need to be added, updating any existing, and removing any that no longer match the payload. Available regions are differentiated by the `available_region_id` property.
      tags:
        - equipment
      requestBody:
        description: >
          The new available regions for the equipment item.
        required: true
        content:
          application/json:
            schema:
              type: array
              items:  
                $ref: '#/components/schemas/EquipmentAvailableRegion'
                  
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:  
                  $ref: '#/components/schemas/ProductFilterOption'
              
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /equipment/${id}/images:
    parameters:
    - name: id
      in: path
      description: the equipment id
      required: true
      schema:
        type: integer
    get:
      summary: gets images for a given equipment item
      tags:
        - equipment
      responses:
        200:
          description: OK
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/EquipmentImage'
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
    post:
      summary: replaces the images for an equipment item
      description: Replaces the imagesfor an equipment item, adding those that need to be added, updating any existing, and removing any that no longer match the payload. Images are differentiated by the `image_link` property.
      tags:
        - equipment
      requestBody:
        description: >
          The current images for the equipment item.
        required: true
        content:
          application/json:
            schema:
              type: array
              items:  
                $ref: '#/components/schemas/EquipmentImage'
                  
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:  
                  $ref: '#/components/schemas/ProductFilterOption'
              
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /equipment/search:
    post:
      summary: Perform an advanced search for equipment.
      description: Extends basic search capabilities, allowing a search term to be used across different equipment fields. Additional filter criteria can be used to further narrow the search-term search.
      tags:
        - equipment
      parameters:
        - in: header
          name: Accept
          required: false
          description: If omitted, `application/json` is assumed.
          schema:
            type: string
            enum:
              - application/json
              - text/csv
      requestBody:
        description: >
          The search criteria to use for the search.
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/SearchCriteria'
              - type: object
                properties:
                  search_term:
                    type: string
                    description: wildcard search value to be used for searching any/all of the `search_term_fields`
                  search_term_fields:
                    description: which equipment fields the search term is applied to
                    type: array
                    items:
                      type: string
                      enum: [model, brand_en, type_en]
                  search_filter:
                    type: string
                    enum: [brand, equipment type]
                    description: >
                      the type of filtering criteria to further refine results.
                  search_filter_value:
                    oneOf:
                      - type: integer
                      - type: string
                    description: >
                      The value to be used in the search filter criteria depends on the
                      value used for the `search_filter`

                        | search_filter | value is a(n) | matching |
                        |---------------|------------|----------|
                        | brand | integer | brand_id of the equipment
                        | equipment type | integer | equipment_type_id of the equipment 
                   
            examples:
              example1:
                description: >
                  Search for equipment having a `model` matching `SureSigns` (wildcard),
                  
                  
                  **and** further apply a filter to get results where equipment has **equipment type** (`equipment_type_id`) = `1`.
                  
                  
                  Order the results by `id` (equipment id). 
                  

                  Do not skip any results, but limit the output to `10` entries.
                value:
                  offset: 0
                  limit: 10
                  order_by:
                    - id
                  search_term: SureSigns
                  search_term_fields:
                    - model
                  category_id:
                    - 7
                  search_filter: equipment type
                  search_filter_value: 1
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    description: The total number of matching equipment in the database.
                    type: integer
                  product_views:
                    type: array
                    items:
                      $ref: '#/components/schemas/EquipmentView'
            text/csv: {} 
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
  /equipment/search/download:
    post:
      summary: Download the results of an advanced search for equipment.
      description: Extends basic search capabilities, allowing a search term to be used across different equipment fields. Additional filter criteria can be used to further narrow the search-term search.
      tags:
        - equipment
      parameters:
        - in: header
          name: Accept
          required: false
          description: If omitted, `application/json` is assumed.
          schema:
            type: string
            enum:
              - application/json
              - text/csv
      requestBody:
        description: >
          The search criteria to use for the search.
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/SearchCriteria'
              - type: object
                properties:
                  search_term:
                    type: string
                    description: wildcard search value to be used for searching any/all of the `search_term_fields`
                  search_term_fields:
                    description: which equipment fields the search term is applied to
                    type: array
                    items:
                      type: string
                      enum: [model, brand_en, type_en]
                  search_filter:
                    type: string
                    enum: [brand, equipment type]
                    description: >
                      the type of filtering criteria to further refine results.
                  search_filter_value:
                    oneOf:
                      - type: integer
                      - type: string
                    description: >
                      The value to be used in the search filter criteria depends on the
                      value used for the `search_filter`

                        | search_filter | value is a(n) | matching |
                        |---------------|------------|----------|
                        | brand | integer | brand_id of the equipment
                        | equipment type | integer | equipment_type_id of the equipment 
                   
            examples:
              example1:
                description: >
                  Search for equipment having a `model` matching `SureSigns` (wildcard),
                  
                  
                  **and** further apply a filter to get results where equipment has **equipment type** (`equipment_type_id`) = `1`.
                  
                  
                  Order the results by `id` (equipment id). 
                  

                  Do not skip any results, but limit the output to `10` entries.
                value:
                  offset: 0
                  limit: 10
                  order_by:
                    - id
                  search_term: SureSigns
                  search_term_fields:
                    - model
                  category_id:
                    - 7
                  search_filter: equipment type
                  search_filter_value: 1
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    description: The total number of matching equipment in the database.
                    type: integer
                  equipment_views:
                    type: array
                    items:
                      $ref: '#/components/schemas/EquipmentView'
                  headers:
                    type: array
                    items:
                      type: object
                      description: a single object whose properties can be used to create headers for CSV/table rendering.
            text/csv: {} 
        400:
          $ref: '#/components/responses/400ClientError'
        500:
          $ref: '#/components/responses/500ServerError'
...
components:
  schemas:
    ApiError:
      type: object
      properties:
        message:
          type: string
          description: A human-readable message describing the error.
        error:
          type: string
          description: The error details. Typically error details may contain sensitive information not intended to be shown directly in the user interface.
    Auditable:
      type: object
      properties:
        created:
          type: string
          format: date-time
          readOnly: true
        updated:
          type: string
          format: date-time
          readOnly: true
        version:
          type: integer
          readOnly: true
...
    Equipment:
      type: object
      description: An equipment model. Equipment represents items that a Product may be compatible with (e.g. can connect to).
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          brand_id:
            description: References the brand.
            type: integer
          equipment_type_id:
            description: References the equipment type.
            type: integer
          model:
            description: Model number/identifier.
            type: string
            maxLength: 255
      - $ref: '#/components/schemas/Auditable'
    EquipmentView:
      type: object
      description: The view of an equipment item that includes equipment type and brand information..
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          model:
            description: Model number/identifier.
            type: string
            maxLength: 255
          equipment_type_id:
            description: References the equipment type.
            type: integer
          type_en:
            description: English equipment type name.
            type: string
            maxLength: 255
          type_zh:
            description: Chinese equipment type name.
            type: string
            maxLength: 255
          brand_id:
            description: References the brand.
            type: integer
          brand_en:
            description: English brand name.
            type: string
            maxLength: 255
          brand_zh:
            description: Chinese brand name.
            type: string
            maxLength: 255
          is_oem:
            description: Whether the referenced brand is an oem.
            type: boolean
      - $ref: '#/components/schemas/Auditable'
    EquipmentAvailableRegion:
      type: object
      description: A geographic region where an equipment item is available.
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          equipment_id:
            description: References the equipment item.
            type: integer
          available_region_id:
            description: References the available region.
            type: integer
      - $ref: '#/components/schemas/Auditable'
    EquipmentAvailableRegionView:
      type: object
      description: A geographic region where an equipment item is available, including extended information.
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          equipment_id:
            description: References the equipment item.
            type: integer
          equipment_model:: 
            type: string
          available_region_id:
            description: References the available region.
            type: integer
          available_region_name_en:
            type: string
      - $ref: '#/components/schemas/Auditable'
    EquipmentImage:
      type: object
      description: An equipment-specific image.
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          equipment_id:
            description: References the equipment item.
            type: integer
          image_link:
            description: The image URL.
            type: string
      - $ref: '#/components/schemas/Auditable'
...
    ProductEquipment:
      type: object
      description: The equipment reference for a product.
      allOf:
      - type: object
        properties:
          id:
            type: integer
          product_id:
            type: integer
          equipment_id:
            type: integer
      - $ref: '#/components/schemas/Auditable'    
    ProductEquipmentView:
      type: object
      description: Equipment reference for a product, including extend information.
      allOf:
      - type: object
        properties:
          id:
            type: integer
            readOnly: true
          product_id:
            type: integer
          equipment_id:
            type: integer
          model:
            description: The model number for the equipment.
            type: string
            maxLength: 255
          brand_en:
            description: The equipment brand name (Chinese)
            type: string
            maxLength: 255
          brand_zh:
            description: The equipment brand name (Chinese)
            type: string
            maxLength: 255
      - $ref: '#/components/schemas/Auditable'
...
    ProductFilterOption:
      type: object
      description: The filter option reference for a product.
      allOf:
      - type: object
        properties:
          id:
            type: integer
          product_id:
            type: integer
          filter_option_id:
            type: integer
          priority_order:
            description: the display order/priority of the option
            type: integer
      - $ref: '#/components/schemas/Auditable'
...
    SearchCriteria:
      type: object
      allOf:
      - type: object
        properties:
          offset:
            type: integer
            description: for pagination. A zero-indexed value indicating where to begin the output with in the set of all possible search results.
          limit:
            type: integer
            description: for pagination. The number of records to return in a single query (up to 1000).
          order_by:
            type: array
            description: Describes the ordering of the search results. Provide the field names to order by. Prefix the field name with a `-` if you want ordering to be in descending order.
            items:
              type: string
          search_term:
            type: string
            description: the search term value. It will be treated as a wildcard search within the `search_term_fields` values.
          search_term_fields:
            type: array
            description: the list of fields you want the `search_term` to apply to. The criteria will be applied as a (OR) when conducting the search across more than one field.
            items:
              type: string        
...
  responses:
    400ClientError:
      description: client error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'
...
    500ServerError:
      description: server error
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/ApiError'